tests/wpt/web-platform-tests/webstorage/storage_session_window_open_second.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebStorage Test: sessionStorage - second page</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
    test(function() {
        assert_true("sessionStorage" in window, "sessionStorage exist");

        var storage = window.sessionStorage;

        assert_equals(storage.getItem("FOO"), "BAR", "storage.getItem('FOO')");
        storage.setItem("FOO", "BAR-NEWWINDOW");
        assert_equals(storage.getItem("FOO"), "BAR-NEWWINDOW", "value for FOO after changing");

        assert_equals(window.opener.sessionStorage.getItem("FOO"), "BAR", "value for FOO in my opening window");

    }, "A new window to make sure there is a copy of the previous window's sessionStorage, and that they diverge after a change");

</script>
</body>
</html>
